style(multi-dim): split into multiple files
